Set up apicalls file and collect character data
